Add files via upload
Problem Statement

The problem statement chosen for this project is to predict fraudulent credit card transactions with the help of machine learning models.
In this project, you will analyze customer-level data collected and analyzed during a research collaboration between Worldline and the Machine Learning Group. 

The dataset is taken from the Kaggle website and has a total of 2,84,807 transactions, of which 492 are fraudulent. Since the dataset is highly imbalanced, so it needs to be handled before model building.

Business Problem Overview
For many banks, retaining high profitable customers is the number one business goal. Banking fraud, however, poses a significant threat to this goal for different banks. In terms of substantial financial losses, trust, and credibility, this is a concerning issue to both banks and customers alike.

It has been estimated by Nilson's report that by 2020 banking frauds would account for $30 billion worldwide. With the rise in digital payment channels, the number of fraudulent transactions is also increasing in different ways. 

In the banking industry, credit card fraud detection using machine learning is not just a trend but a necessity for them to put proactive monitoring and fraud prevention mechanisms in place. Machine learning is helping these institutions to reduce time-consuming manual reviews, costly chargebacks and fees, and denials of legitimate transactions. 

Understanding and Defining Fraud
Credit card fraud is any dishonest act and behavior to obtain information without the proper authorization from the account holder for financial gain. Among different ways of fraud, Skimming is the most common one, which is the way of duplicating information located on the magnetic stripe of the card.  Apart from this, the other ways are:
Manipulation/alteration of genuine cards
Creation of counterfeit cards
Stolen/lost credit cards
Fraudulent telemarketing

Data Dictionary
The dataset can be downloaded using this link: https://www.kaggle.com/datasets/mlg-ulb/creditcardfraud 

The data set includes credit card transactions made by European cardholders over two days in September 2013. Out of a total of 2,84,807 transactions, 492 were fraudulent. This data set is highly unbalanced, with the positive class (frauds) accounting for 0.172% of the total transactions. The data set has also been modified with Principal Component Analysis (PCA) to maintain confidentiality. Apart from ‘time’ and ‘amount’, all the other features (V1, V2, V3, up to V28) are the principal components obtained using PCA. The feature 'time' contains the seconds elapsed between the first transaction in the data set and the subsequent transactions. The feature 'amount' is the transaction amount. The feature 'class' represents class labeling, and it takes the value 1 in cases of fraud and 0 in others.